Update and rename __main__.py to keyboard_main.py
